GPUCP-606: Use https for title urls
The UCP sets wgServer to different values depending on the originating
request protocol; code executed in a Job context will have http:// in
the wgServer url.  This means Title::getFullURL varies in result
depending on whether the code runs in a Job or on the web.

Force HTTPS everywhere because it's <current year>.