Vendor Reverb client code
We currently use the hydrawiki/reverb-client composer package to provide
us a Reverb client, which is still hosted on GitLab and is an useless
layer of indirection. This complicates the official upgrade to PHP 8.0
as some of its dependencies refuse to install on PHP 8.0.

So, move the client code and tests to this repo, and add an action that
runs client tests. Also remove the hydrawiki codesniffer dependency
because it's unmaintained and likewise refuses to install on PHP 8.0.